add conftest

+++ conftest.py
@@ -58,7 +58,8 @@ def check_std_out_err(capfd):
 def pytest_sessionfinish(session, exitstatus):
     if exitstatus == NO_TESTS_COLLECTED:
         session.exitstatus = SUCCESS
-    os.remove("demistomock.py")
+    if os.path.exists("demistomock.py"):
+        os.remove("demistomock.py")
     shutil.rmtree("test_data", ignore_errors=True)
 
 def pytest_configure(config):
